separated interators
