+++ app/src/main/java/com/example/searchmovieapp/domain/interactors/MovieDetailsInteractor.kt
@@ -1,0 +1,8 @@
+package com.example.searchmovieapp.domain.interactors
+
+import com.example.searchmovieapp.data.remote.entities.MovieDetailsEntity
+import com.example.searchmovieapp.domain.data.ResultWrapper
+
+interface MovieDetailsInteractor {
+    suspend fun getMovieDetails(movieId: Int): ResultWrapper<MovieDetailsEntity>
+}
